Jarvis/requirements.txt
python-dotenv==1.0.0
spotipy==2.23.0

# Vocal Assistant
SpeechRecognition==3.10.3
pyaudio==0.2.14
pyttsx3==2.9
soundfile==0.12.1
openai==0.27.6
groq==0.5.0